Adding superclass for authorizable services
